+++ tests/test_default_qubit.py
@@ -96,267 +96,6 @@ def prep_par(par, op):
 def include_inverses_with_test_data(test_data):
     return test_data + [(item[0] + ".inv", item[1], item[2]) for item in test_data]
 
-class TestAuxillaryFunctions:
-    """Test auxillary functions."""
-
-    def test_spectral_decomposition(self, tol):
-        """Test that the correct spectral decomposition is returned."""
-
-        a, P = spectral_decomposition(H)
-
-        # verify that H = \sum_k a_k P_k
-        assert np.allclose(H, np.einsum("i,ijk->jk", a, P), atol=tol, rtol=0)
-
-    def test_phase_shift(self, tol):
-        """Test phase shift is correct"""
-
-        # test identity for theta=0
-        assert np.allclose(Rphi(0), np.identity(2), atol=tol, rtol=0)
-
-        # test arbitrary phase shift
-        phi = 0.5432
-        expected = np.array([[1, 0], [0, np.exp(1j * phi)]])
-        assert np.allclose(Rphi(phi), expected, atol=tol, rtol=0)
-
-    def test_x_rotation(self, tol):
-        """Test x rotation is correct"""
-
-        # test identity for theta=0
-        assert np.allclose(Rotx(0), np.identity(2), atol=tol, rtol=0)
-
-        # test identity for theta=pi/2
-        expected = np.array([[1, -1j], [-1j, 1]]) / np.sqrt(2)
-        assert np.allclose(Rotx(np.pi / 2), expected, atol=tol, rtol=0)
-
-        # test identity for theta=pi
-        expected = -1j * np.array([[0, 1], [1, 0]])
-        assert np.allclose(Rotx(np.pi), expected, atol=tol, rtol=0)
-
-    def test_y_rotation(self, tol):
-        """Test y rotation is correct"""
-
-        # test identity for theta=0
-        assert np.allclose(Roty(0), np.identity(2), atol=tol, rtol=0)
-
-        # test identity for theta=pi/2
-        expected = np.array([[1, -1], [1, 1]]) / np.sqrt(2)
-        assert np.allclose(Roty(np.pi / 2), expected, atol=tol, rtol=0)
-
-        # test identity for theta=pi
-        expected = np.array([[0, -1], [1, 0]])
-        assert np.allclose(Roty(np.pi), expected, atol=tol, rtol=0)
-
-    def test_z_rotation(self, tol):
-        """Test z rotation is correct"""
-
-        # test identity for theta=0
-        assert np.allclose(Rotz(0), np.identity(2), atol=tol, rtol=0)
-
-        # test identity for theta=pi/2
-        expected = np.diag(np.exp([-1j * np.pi / 4, 1j * np.pi / 4]))
-        assert np.allclose(Rotz(np.pi / 2), expected, atol=tol, rtol=0)
-
-        # test identity for theta=pi
-        assert np.allclose(Rotz(np.pi), -1j * Z, atol=tol, rtol=0)
-
-    def test_arbitrary_rotation(self, tol):
-        """Test arbitrary single qubit rotation is correct"""
-
-        # test identity for phi,theta,omega=0
-        assert np.allclose(Rot3(0, 0, 0), np.identity(2), atol=tol, rtol=0)
-
-        # expected result
-        def arbitrary_rotation(x, y, z):
-            """arbitrary single qubit rotation"""
-            c = np.cos(y / 2)
-            s = np.sin(y / 2)
-            return np.array(
-                [
-                    [np.exp(-0.5j * (x + z)) * c, -np.exp(0.5j * (x - z)) * s],
-                    [np.exp(-0.5j * (x - z)) * s, np.exp(0.5j * (x + z)) * c],
-                ]
-            )
-
-        a, b, c = 0.432, -0.152, 0.9234
-        assert np.allclose(Rot3(a, b, c), arbitrary_rotation(a, b, c), atol=tol, rtol=0)
-
-    def test_C_x_rotation(self, tol):
-        """Test controlled x rotation is correct"""
-
-        # test identity for theta=0
-        assert np.allclose(CRotx(0), np.identity(4), atol=tol, rtol=0)
-
-        # test identity for theta=pi/2
-        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1/np.sqrt(2), -1j/np.sqrt(2)], [0, 0, -1j/np.sqrt(2), 1/np.sqrt(2)]])
-        assert np.allclose(CRotx(np.pi / 2), expected, atol=tol, rtol=0)
-
-        # test identity for theta=pi
-        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1j], [0, 0, -1j, 0]])
-        assert np.allclose(CRotx(np.pi), expected, atol=tol, rtol=0)
-
-    def test_C_y_rotation(self, tol):
-        """Test controlled y rotation is correct"""
-
-        # test identity for theta=0
-        assert np.allclose(CRoty(0), np.identity(4), atol=tol, rtol=0)
-
-        # test identity for theta=pi/2
-        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1/np.sqrt(2), -1/np.sqrt(2)], [0, 0, 1/np.sqrt(2), 1/np.sqrt(2)]])
-        assert np.allclose(CRoty(np.pi / 2), expected, atol=tol, rtol=0)
-
-        # test identity for theta=pi
-        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]])
-        assert np.allclose(CRoty(np.pi), expected, atol=tol, rtol=0)
-
-    def test_C_z_rotation(self, tol):
-        """Test controlled z rotation is correct"""
-
-        # test identity for theta=0
-        assert np.allclose(CRotz(0), np.identity(4), atol=tol, rtol=0)
-
-        # test identity for theta=pi/2
-        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, np.exp(-1j * np.pi / 4), 0], [0, 0, 0, np.exp(1j * np.pi / 4)]])
-        assert np.allclose(CRotz(np.pi / 2), expected, atol=tol, rtol=0)
-
-        # test identity for theta=pi
-        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, -1j, 0], [0, 0, 0, 1j]])
-        assert np.allclose(CRotz(np.pi), expected, atol=tol, rtol=0)
-
-    def test_controlled_arbitrary_rotation(self, tol):
-        """Test controlled arbitrary rotation is correct"""
-
-        # test identity for phi,theta,omega=0
-        assert np.allclose(CRot3(0, 0, 0), np.identity(4), atol=tol, rtol=0)
-
-        # test identity for phi,theta,omega=pi
-        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]])
-        assert np.allclose(CRot3(np.pi, np.pi, np.pi), expected, atol=tol, rtol=0)
-
-        def arbitrary_Crotation(x, y, z):
-            """controlled arbitrary single qubit rotation"""
-            c = np.cos(y / 2)
-            s = np.sin(y / 2)
-            return np.array(
-                [
-                    [1, 0, 0, 0],
-                    [0, 1, 0, 0],
-                    [0, 0, np.exp(-0.5j * (x + z)) * c, -np.exp(0.5j * (x - z)) * s],
-                    [0, 0, np.exp(-0.5j * (x - z)) * s, np.exp(0.5j * (x + z)) * c]
-                ]
-            )
-
-        a, b, c = 0.432, -0.152, 0.9234
-        assert np.allclose(CRot3(a, b, c), arbitrary_Crotation(a, b, c), atol=tol, rtol=0)
-
-class TestStateFunctions:
-    """Arbitrary state and operator tests."""
-
-    def test_unitary(self, tol):
-        """Test that the unitary function produces the correct output."""
-
-        out = unitary(U)
-
-        # verify output type
-        assert isinstance(out, np.ndarray)
-
-        # verify equivalent to input state
-        assert np.allclose(out, U, atol=tol, rtol=0)
-
-    def test_unitary_exceptions(self):
-        """Tests that the unitary function raises the proper errors."""
-
-        # test non-square matrix
-        with pytest.raises(ValueError, match="must be a square matrix"):
-            unitary(U[1:])
-
-        # test non-unitary matrix
-        U3 = U.copy()
-        U3[0, 0] += 0.5
-        with pytest.raises(ValueError, match="must be unitary"):
-            unitary(U3)
-
-    def test_hermitian(self, tol):
-        """Test that the hermitian function produces the correct output."""
-
-        out = hermitian(H)
-
-        # verify output type
-        assert isinstance(out, np.ndarray)
-
-        # verify equivalent to input state
-        assert np.allclose(out, H, atol=tol, rtol=0)
-
-    def test_hermitian_exceptions(self):
-        """Tests that the hermitian function raises the proper errors."""
-
-        # test non-square matrix
-        with pytest.raises(ValueError, match="must be a square matrix"):
-            hermitian(H[1:])
-
-        # test non-Hermitian matrix
-        H2 = H.copy()
-        H2[0, 1] = H2[0, 1].conj()
-        with pytest.raises(ValueError, match="must be Hermitian"):
-            hermitian(H2)
-
-class TestOperatorMatrices:
-    """Tests that get_operator_matrix returns the correct matrix."""
-
-    @pytest.mark.parametrize("name,expected", [
-        ("PauliX", np.array([[0, 1], [1, 0]])),
-        ("PauliY", np.array([[0, -1j], [1j, 0]])),
-        ("PauliZ", np.array([[1, 0], [0, -1]])),
-        ("S", np.array([[1, 0], [0, 1j]])),
-        ("T", np.array([[1, 0], [0, np.exp(1j * np.pi / 4)]])),
-        ("Hadamard", np.array([[1, 1], [1, -1]])/np.sqrt(2)),
-        ("CNOT", np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]])),
-        ("SWAP", np.array([[1, 0, 0, 0], [0, 0, 1, 0], [0, 1, 0, 0], [0, 0, 0, 1]])),
-        ("CSWAP",np.array([[1, 0, 0, 0, 0, 0, 0, 0],
-                           [0, 1, 0, 0, 0, 0, 0, 0],
-                           [0, 0, 1, 0, 0, 0, 0, 0],
-                           [0, 0, 0, 1, 0, 0, 0, 0],
-                           [0, 0, 0, 0, 1, 0, 0, 0],
-                           [0, 0, 0, 0, 0, 0, 1, 0],
-                           [0, 0, 0, 0, 0, 1, 0, 0],
-                           [0, 0, 0, 0, 0, 0, 0, 1]])),
-        ("CZ", np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, -1]])),
-    ])
-    def test_get_operator_matrix_no_parameters(self, qubit_device_3_wires, tol, name, expected):
-        """Tests that get_operator_matrix returns the correct matrix."""
-
-        res = qubit_device_3_wires._get_operator_matrix(name, ())
-
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-    @pytest.mark.parametrize("name,expected,par", [
-        ('PhaseShift', lambda phi: np.array([[1, 0], [0, np.exp(1j*phi)]]), [0.223]),
-        ('RX', lambda phi: np.array([[math.cos(phi/2), -1j*math.sin(phi/2)], [-1j*math.sin(phi/2), math.cos(phi/2)]]), [0.223]),
-        ('RY', lambda phi: np.array([[math.cos(phi/2), -math.sin(phi/2)], [math.sin(phi/2), math.cos(phi/2)]]), [0.223]),
-        ('RZ', lambda phi: np.array([[cmath.exp(-1j*phi/2), 0], [0, cmath.exp(1j*phi/2)]]), [0.223]),
-        ('Rot', lambda phi, theta, omega: np.array([[cmath.exp(-1j*(phi+omega)/2)*math.cos(theta/2), -cmath.exp(1j*(phi-omega)/2)*math.sin(theta/2)], [cmath.exp(-1j*(phi-omega)/2)*math.sin(theta/2), cmath.exp(1j*(phi+omega)/2)*math.cos(theta/2)]]), [0.223, 0.153, 1.212]),
-        ('CRX', lambda phi: np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, math.cos(phi/2), -1j*math.sin(phi/2)], [0, 0, -1j*math.sin(phi/2), math.cos(phi/2)]]), [0.223]),
-        ('CRY', lambda phi: np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, math.cos(phi/2), -math.sin(phi/2)], [0, 0, math.sin(phi/2), math.cos(phi/2)]]), [0.223]),
-        ('CRZ', lambda phi: np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, cmath.exp(-1j*phi/2), 0], [0, 0, 0, cmath.exp(1j*phi/2)]]), [0.223]),
-        ('CRot', lambda phi, theta, omega: np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, cmath.exp(-1j*(phi+omega)/2)*math.cos(theta/2), -cmath.exp(1j*(phi-omega)/2)*math.sin(theta/2)], [0, 0, cmath.exp(-1j*(phi-omega)/2)*math.sin(theta/2), cmath.exp(1j*(phi+omega)/2)*math.cos(theta/2)]]), [0.223, 0.153, 1.212]),
-        ('QubitUnitary', lambda U: np.asarray(U), [np.array([[0.83645892 - 0.40533293j, -0.20215326 + 0.30850569j], [-0.23889780 - 0.28101519j, -0.88031770 - 0.29832709j]])]),
-        ('Hermitian', lambda H: np.asarray(H), [np.array([[1.02789352, 1.61296440 - 0.3498192j], [1.61296440 + 0.3498192j, 1.23920938 + 0j]])]),
-        # Identity will always return a 2x2 Identity, but is still parameterized
-        ('Identity', lambda n: np.eye(2), [2])
-    ])
-    def test_get_operator_matrix_with_parameters(self, qubit_device_2_wires, tol, name, expected, par):
-        """Tests that get_operator_matrix returns the correct matrix building functions."""
-
-        res = qubit_device_2_wires._get_operator_matrix(name, par)
-
-        assert np.allclose(res, expected(*par), atol=tol, rtol=0)
-
-    @pytest.mark.parametrize("name", ["BasisState", "QubitStateVector"])
-    def test_get_operator_matrix_none(self, qubit_device_2_wires, name):
-        """Tests that get_operator_matrix returns none for direct state manipulations."""
-
-        res = qubit_device_2_wires._get_operator_matrix(name, ())
-
-        assert res is None
 
 class TestApply:
     """Tests that operations and inverses of certain operations are applied correctly or that the proper
