+++ tests/test_operation.py
@@ -23,7 +23,7 @@ import numpy as np
 from numpy.linalg import multi_dot
 
 import pennylane as qml
-from pennylane.plugins.default_qubit import I, X, Y, Rotx, Roty, Rotz, CRotx, CRoty, CRotz, CNOT, Rot3, Rphi
+from pennylane.beta.plugins.expt_tensornet import I, X, Y, Rotx, Roty, Rotz, CRotx, CRoty, CRotz, CNOT, Rot3, Rphi
 from pennylane.operation import Tensor
 
 # pylint: disable=no-self-use, no-member, protected-access, pointless-statement
